Use Flatlist for Suggestions and fix tests
