+++ tests/unit/components/Suggestions/Suggestions.test.js
@@ -1,8 +1,8 @@
-import { faker } from "@faker-js/faker";
 import {
+  fireEvent,
   screen, waitFor
 } from "@testing-library/react-native";
-import SuggestionsContainer from "components/Suggestions/SuggestionsContainer";
+import Suggestions from "components/Suggestions/Suggestions";
 import initI18next from "i18n/initI18next";
 import i18next from "i18next";
 import React from "react";
@@ -11,57 +11,13 @@ import useStore from "stores/useStore";
 import factory from "../../../factory";
 import { renderComponent } from "../../../helpers/render";
 
-const mockTaxon = factory( "RemoteTaxon", {
-  name: faker.person.firstName( ),
-  preferred_common_name: faker.person.fullName( ),
-  default_photo: {
-    square_url: faker.image.url( )
-  }
-} );
-
-const mockSuggestionsList = [{
-  combined_score: 90.34,
-  taxon: {
-    ...mockTaxon,
-    id: faker.number.int( )
-  }
-}, {
-  combined_score: 30.32,
-  taxon: {
-    ...mockTaxon,
-    id: faker.number.int( )
-  }
-}];
-
-jest.mock( "sharedHooks/useAuthenticatedQuery", () => ( {
-  __esModule: true,
-  default: () => ( {
-    data: mockSuggestionsList,
-    isLoading: false
-  } )
-} ) );
-
-const mockOfflinePrediction = {
-  score: 0.97363,
-  taxon: {
-    rank_level: 10,
-    name: "Felis Catus",
-    id: 118552
-  }
-};
-
-jest.mock( "components/Suggestions/hooks/useOfflineSuggestions", () => ( {
-  __esModule: true,
-  default: ( ) => ( {
-    offlineSuggestions: [mockOfflinePrediction],
-    loadingOfflineSuggestions: false
-  } )
-} ) );
-
-// Mock api call to observations
-jest.mock( "inaturalistjs" );
+const mockCreateId = jest.fn( );
 
 const initialStoreState = useStore.getState( );
+
+const mockSuggestionsList = [{
+  taxon: factory( "RemoteTaxon" )
+}];
 
 describe( "Suggestions", ( ) => {
   beforeAll( async ( ) => {
@@ -70,40 +26,68 @@ describe( "Suggestions", ( ) => {
   } );
 
   test( "should not have accessibility errors", async ( ) => {
-    renderComponent( <SuggestionsContainer /> );
+    renderComponent( <Suggestions
+      nearbySuggestions={mockSuggestionsList}
+    /> );
 
     const suggestions = await screen.findByTestId( "suggestions" );
     expect( suggestions ).toBeAccessible( );
   } );
 
   it( "should fetch offline suggestions for current photo", async ( ) => {
-    renderComponent( <SuggestionsContainer /> );
+    renderComponent( <Suggestions
+      nearbySuggestions={mockSuggestionsList}
+    /> );
     const taxonTopResult = screen.getByTestId(
-      `SuggestionsList.taxa.${mockOfflinePrediction.taxon.id}`
+      `SuggestionsList.taxa.${mockSuggestionsList[0].taxon.id}`
     );
     await waitFor( ( ) => {
       expect( taxonTopResult ).toBeVisible( );
     } );
   } );
 
-  // it( "should fetch nearby suggestions for current photo", async ( ) => {
-  //   renderComponent( <SuggestionsContainer /> );
-  //   const taxonTopResult = screen.getByTestId(
-  //     `SuggestionsList.taxa.${mockSuggestionsList[0].taxon.id}`
-  //   );
-  //   await waitFor( ( ) => {
-  //     expect( taxonTopResult ).toBeVisible( );
-  //   } );
-  // } );
-
   it( "shows comment section if observation has comment", ( ) => {
     useStore.setState( {
       comment: "This is a test comment"
     } );
-    renderComponent( <SuggestionsContainer /> );
+    renderComponent( <Suggestions
+      nearbySuggestions={mockSuggestionsList}
+    /> );
     const commentSection = screen.getByText(
       i18next.t( "Your-identification-will-be-posted-with-the-following-comment" )
     );
     expect( commentSection ).toBeVisible( );
+  } );
+
+  it( "should display empty text if no suggestions are found", ( ) => {
+    renderComponent( <Suggestions
+      nearbySuggestions={[]}
+    /> );
+    const emptyText = i18next
+      .t( "iNaturalist-isnt-able-to-provide-a-top-ID-suggestion-for-this-photo" );
+    expect( screen.getByText( emptyText ) ).toBeVisible( );
+  } );
+
+  it( "should display a loading wheel if suggestions are loading", ( ) => {
+    renderComponent( <Suggestions
+      nearbySuggestions={[]}
+      loadingSuggestions
+    /> );
+    const loading = screen.getByTestId( "SuggestionsList.loading" );
+    expect( loading ).toBeVisible( );
+  } );
+
+  it( "should create an id when checkmark is pressed", async ( ) => {
+    renderComponent( <Suggestions
+      nearbySuggestions={mockSuggestionsList}
+      onTaxonChosen={mockCreateId}
+    /> );
+    const testID = `SuggestionsList.taxa.${mockSuggestionsList[0].taxon.id}`;
+    const checkmark = screen.getByTestId( `${testID}.checkmark` );
+    expect( checkmark ).toBeVisible( );
+    fireEvent.press( checkmark );
+    await waitFor( ( ) => {
+      expect( mockCreateId ).toHaveBeenCalled( );
+    } );
   } );
 } );
